backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.1.12
chromadb==0.4.22
torch==2.2.1
transformers==4.38.2
accelerate==0.27.2
sentence-transformers==2.5.1
bitsandbytes==0.42.0
protobuf==3.20.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.2.0